Get User Chats with socket.io
