+++ client/src/components/hooks/useGetChatMessages.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { getChatMessages } from '../../api/chats';
 
 export default function useGetChatMessages(_id) {
   const [chatMessages, setChatMessages] = React.useState([]);
@@ -15,18 +14,3 @@ export default function useGetChatMessages(_id) {
 
   return chatMessages;
 }
-
-// export default function useGetChatMessages(_id) {
-//   const [chatMessages, setChatMessages] = React.useState([]);
-
-//   React.useEffect(() => {
-//     const interval = setInterval(() => {
-//       getChatMessages(_id).then(fetchedChatMessages => {
-//         setChatMessages(fetchedChatMessages);
-//       });
-//     }, 2000);
-//     return () => clearInterval(interval);
-//   }, [_id]);
-
-//   return chatMessages;
-// }
