+++ client/src/components/hooks/useGetUserChats.js
@@ -1,12 +1,14 @@
 import React from 'react';
-import { getUserChats } from '../../api/chats';
 
 export default function useGetUserChats(userName) {
   const [userChats, setUserChats] = React.useState([]);
 
   React.useEffect(() => {
-    getUserChats(userName).then(fetchedUserChats => {
-      setUserChats(fetchedUserChats);
+    const io = require('socket.io-client');
+    const socket = io('http://localhost:9090');
+    socket.emit('get-userName', userName);
+    socket.on('user-chats', data => {
+      setUserChats(data);
     });
   }, [userName]);
