+++ lib/chats.js
@@ -1,24 +1,36 @@
 const express = require('express');
 const router = express.Router();
 const mongoose = require('mongoose');
+const io = require('socket.io')(9090, {
+  handlePreflightRequest: (req, res) => {
+    const headers = {
+      'Access-Control-Allow-Headers': 'Content-Type, Authorization',
+      'Access-Control-Allow-Origin': req.headers.origin, //or the specific origin you want to give access to,
+      'Access-Control-Allow-Credentials': true
+    };
+    res.writeHead(200, headers);
+    res.end();
+  }
+});
 
 //Models imports
 const Chat = require('./models/chat');
 
 //Get user-specific chats
-router.get('/userChats/:name', async (request, response) => {
-  try {
-    const chats = await Chat.find({
-      $or: [
-        { $and: [{ userName1: request.params.name }] },
-        { $and: [{ userName2: request.params.name }] }
-      ]
-    });
-    response.json(chats);
-  } catch (error) {
-    console.log(error);
-    response.status(500).json({ message: error.message });
-  }
+io.on('connection', socket => {
+  socket.on('get-userName', userName => {
+    Chat.find(
+      {
+        $or: [{ $and: [{ userName1: userName }] }, { $and: [{ userName2: userName }] }]
+      },
+      function(error, result) {
+        if (error) {
+          throw error;
+        }
+        socket.emit('user-chats', result);
+      }
+    );
+  });
 });
 
 //Get one chat
@@ -33,8 +45,17 @@ router.get('/:id', async (request, response) => {
 });
 
 //Get messages of one chat
-router.get('/:id/messages', getChat, async (request, response) => {
-  response.send(response.chat.messages);
+io.on('connection', socket => {
+  socket.on('send-chat-id', _id => {
+    const ObjectId = require('mongodb').ObjectID;
+    const chatId = new ObjectId(_id);
+    Chat.findOne(chatId, function(error, result) {
+      if (error) {
+        throw error;
+      }
+      socket.emit('chat-messages', result);
+    });
+  });
 });
 
 //Get chat ID by userName1 and userName2
